Simplify content provider URIs
and remove the variant to get a single user only

+++ android_app/app/src/main/java/com/health/openscale/core/database/ScaleMeasurementDAO.java
@@ -68,5 +68,5 @@ public interface ScaleMeasurementDAO {
 
     // selectAll() is equivalent to getAll(), but returns a Cursor, for exposing via a ContentProvider.
     @Query("SELECT id as _ID, datetime, weight, fat, water, muscle FROM scaleMeasurements WHERE userId = :userId AND enabled = 1 ORDER BY datetime DESC")
-    Cursor selectAll(int userId);
+    Cursor selectAll(long userId);
 }
